Added C-style API.

+++ src/lib/takamori/streams/CFileStream.h
@@ -21,7 +21,7 @@ CGSS_NS_BEGIN
 
         virtual uint32_t Read(void *buffer, uint32_t bufferSize, size_t offset, uint32_t count) override;
 
-        virtual uint32_t Write(const void *buffer, uint32_t bufferSize, uint32_t offset, uint32_t count) override;
+        virtual uint32_t Write(const void *buffer, uint32_t bufferSize, size_t offset, uint32_t count) override;
 
         virtual bool_t IsWritable() const override;
